# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CreateVehicleRequest:
    properties:
      brand:
        type: string
      capacity:
        type: integer
      chassis_num:
        type: string
      color:
        type: string
      insurance_exp:
        description: RFC3339
        type: string
      model:
        type: string
      plate:
        type: string
      tech_review_exp:
        description: RFC3339
        type: string
      vehicle_type:
        type: string
      year:
        type: integer
    required:
    - brand
    - model
    - plate
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.DriverResponse:
    properties:
      id:
        type: string
      license_expiry:
        type: string
      license_number:
        type: string
      license_type:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.GuardianResponse:
    properties:
      id:
        type: string
      is_primary:
        type: boolean
      relation:
        type: string
      student_id:
        type: string
      user_id:
        type: string
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LinkGuardianRequest:
    properties:
      is_primary:
        type: boolean
      relation:
        type: string
      student_id:
        type: string
      user_id:
        type: string
    required:
    - relation
    - student_id
    - user_id
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListDriversResponse:
    properties:
      drivers:
        items:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.DriverResponse'
        type: array
      total:
        type: integer
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListVehiclesResponse:
    properties:
      total:
        type: integer
      vehicles:
        items:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.VehicleResponse'
        type: array
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.RegisterDriverRequest:
    properties:
      license_expiry:
        description: RFC3339
        type: string
      license_number:
        type: string
      license_type:
        type: string
      user_id:
        type: string
    required:
    - license_number
    - user_id
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.UpdateDriverRequest:
    properties:
      license_expiry:
        type: string
      license_number:
        type: string
      license_type:
        type: string
      status:
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.VehicleResponse:
    properties:
      brand:
        type: string
      capacity:
        type: integer
      chassis_num:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      insurance_exp:
        type: string
      model:
        type: string
      plate:
        type: string
      status:
        type: string
      tech_review_exp:
        type: string
      updated_at:
        type: string
      vehicle_type:
        type: string
      year:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for the gateway service of the School Tracking System.
  title: Gateway Service API
  version: "1.0"
paths:
...
  /fleet/drivers:
    get:
      description: Retrieve a paginated list of drivers
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListDriversResponse'
      security:
      - BearerAuth: []
      summary: List drivers
      tags:
      - fleet
    post:
      consumes:
      - application/json
      description: Register a new driver linked to an Auth service user
      parameters:
      - description: Register Driver Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.RegisterDriverRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.DriverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a driver
      tags:
      - fleet
  /fleet/drivers/{id}:
    get:
      description: Retrieve driver details
      parameters:
      - description: Driver UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.DriverResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a driver by ID
      tags:
      - fleet
    put:
      consumes:
      - application/json
      description: Modify driver details
      parameters:
      - description: Driver UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update Driver Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.UpdateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.DriverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a driver
      tags:
      - fleet
  /fleet/guardians:
    post:
      consumes:
      - application/json
      description: Create a relationship between a user (guardian role) and a student
      parameters:
      - description: Link Guardian Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LinkGuardianRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.GuardianResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Link a guardian to a student
      tags:
      - fleet
  /fleet/guardians/{id}:
    delete:
      description: Remove the relationship between a guardian and a student
      parameters:
      - description: Guardian Relationship UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlink a guardian
      tags:
      - fleet
...
  /fleet/students/{student_id}/guardians:
    get:
      description: List all guardians linked to a specific student
      parameters:
      - description: Student UUID
        in: path
        name: student_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.GuardianResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get student guardians
      tags:
      - fleet
  /fleet/vehicles:
    get:
      description: Retrieve a paginated list of vehicles in the fleet
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListVehiclesResponse'
      security:
      - BearerAuth: []
      summary: List all vehicles
      tags:
      - fleet
    post:
      consumes:
      - application/json
      description: Register a new vehicle in the fleet
      parameters:
      - description: Create Vehicle Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.VehicleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new vehicle
      tags:
      - fleet
  /fleet/vehicles/{id}:
    get:
      description: Retrieve a single vehicle by its UUID
      parameters:
      - description: Vehicle UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.VehicleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a vehicle by ID
      tags:
      - fleet
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"